move workshop utility function into workshop.util and added coverage
