+++ server/src/workshop/workshop-util.ts
@@ -1,13 +1,17 @@
+import { isWorkshopUUID } from "@tsconline/shared";
+import { isUserInWorkshopAndWorkshopIsActive } from "./database";
+import { OperationResult } from "./types";
+
 /**
  * Extracts the workshop ID from a workshop UUID
  * @param uuid
  * @returns
  */
 export function getWorkshopIdFromUUID(uuid: string): number | null {
-  const workshopId = parseInt(uuid.split("-")[1] ?? "");
-  if (isNaN(workshopId)) {
+  if (!isWorkshopUUID(uuid)) {
     return null;
   }
+  const workshopId = parseInt(uuid.split("-")[1] ?? "");
   return workshopId;
 }
 
@@ -35,4 +39,20 @@ export function getWorkshopUUIDFromWorkshopId(workshopId: number): string {
  */
 export function isUUIDFolderAWorkshopFolder(folderName: string): boolean {
   return folderName.startsWith("workshop-") && getWorkshopIdFromUUID(folderName) !== null;
+}
+/**
+ * Checks to make sure the user is in the workshop and the workshop is active
+ * @param workshopUUID the workshop's UUID
+ * @param userId the user's userId
+ * @returns
+ */
+export function verifyWorkshopValidity(workshopUUID: string, userId: number): OperationResult {
+  const workshopId = getWorkshopIdFromUUID(workshopUUID);
+  if (!workshopId) {
+    return { code: 400, message: "Invalid workshop UUID" };
+  }
+  if (!isUserInWorkshopAndWorkshopIsActive(workshopId, userId)) {
+    return { code: 401, message: "User does not have access to this workshop" };
+  }
+  return { code: 200, message: "Success" };
 }
